Database/requirements.txt
Flask == 2.3.2
requests == 2.31.0
pymongo == 4.5.0
python-dotenv == 1.0.0